# apps/rust-server/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InstantML Rust API",
    "description": "Current Rust/ClickHouse API. See docs/architecture/current-api.md for examples and operational notes.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/clerk": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "auth_clerk",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClerkAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session payload (with optional onboarding API key); HttpOnly session cookie is set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionWithOnboardingKey"
                }
              }
            }
          },
          "401": {
            "description": "Clerk auth not configured or token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Origin not allowed or signup restricted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/auth/dev/google": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "auth_dev_google",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevGoogleAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session payload; HttpOnly session cookie is set on the response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionPayload"
                }
              }
            }
          },
          "401": {
            "description": "Dev auth is not enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/auth/device-code/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "device_code_confirm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCodeConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Grant confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCodeConfirmResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "user_code not found or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
    "/api/auth/device-code/poll": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "device_code_poll",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCodePollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "status: pending | authorized | denied | expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCodePollResponse"
                }
              }
            }
          },
          "429": {
            "description": "slow_down: polling too fast",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/auth/device-code/start": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "device_code_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCodeStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Device-code grant initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCodeStartResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "auth_logout",
        "responses": {
          "200": {
            "description": "Session revoked and cookie cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionUnauthenticated"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
    "/api/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "auth_session",
        "responses": {
          "200": {
            "description": "Unauthenticated session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionUnauthenticated"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
    "/api/auth/switch-organization": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Re-point the caller's session at a different org they belong to.",
        "description": "Used by the dashboard org-switcher. The session token is unchanged — only\nthe bound `org_id` on the session row is updated, so subsequent requests\nfrom the same cookie scope to the newly selected org.",
        "operationId": "auth_switch_organization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Refreshed session payload bound to the target org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionPayload"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "No active membership in target organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthSessionPayload": {
        "type": "object",
        "required": [
          "authenticated",
          "session",
          "user",
          "organization",
          "membership",
          "memberships",
          "account_type"
        ],
        "properties": {
          "account_type": {
            "type": "string"
          },
          "authenticated": {
            "type": "boolean"
          },
          "billing_checkout": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BillingCheckoutInfo"
              }
            ]
          },
          "membership": {
            "$ref": "#/components/schemas/MembershipRow"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MembershipRow"
            }
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRow"
          },
          "provisioning": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProvisioningStatusPayload"
              }
            ]
          },
          "session": {
            "$ref": "#/components/schemas/UserSessionRow"
          },
          "user": {
            "$ref": "#/components/schemas/UserRow"
          }
        }
      },
      "AuthSessionUnauthenticated": {
        "type": "object",
        "required": [
          "authenticated"
        ],
        "properties": {
          "authenticated": {
            "type": "boolean"
          }
        }
      },
      "AuthSessionWithOnboardingKey": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AuthSessionPayload"
          },
          {
            "type": "object",
            "properties": {
              "onboarding_api_key": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/OnboardingApiKey"
                  }
                ]
              }
            }
          }
        ],
        "description": "Wrapper for `auth_dev_google` / `auth_clerk` responses, which serialize\n`AuthSessionPayload` plus an optional `onboarding_api_key` produced for\nbrand-new accounts."
      },
...
      "BillingCheckoutInfo": {
        "type": "object",
        "required": [
          "intent_id",
          "status"
        ],
        "properties": {
          "intent_id": {
            "type": "string",
            "format": "uuid"
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ClerkAuthRequest": {
        "type": "object",
        "properties": {
          "accept_invite_org_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "accept_invite_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "account_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "mode": {
            "type": [
              "string",
              "null"
            ]
          },
          "org_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "plan_tier": {
            "type": [
              "string",
              "null"
            ]
          },
          "seat_emails": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "token": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "DevGoogleAuthRequest": {
        "type": "object",
        "properties": {
          "accept_invite_org_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "accept_invite_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "account_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "mode": {
            "type": [
              "string",
              "null"
            ]
          },
          "org_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "plan_tier": {
            "type": [
              "string",
              "null"
            ]
          },
          "seat_emails": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "DeviceCodeConfirmRequest": {
        "type": "object",
        "properties": {
          "user_code": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "DeviceCodeConfirmResponse": {
        "type": "object",
        "required": [
          "confirmed"
        ],
        "properties": {
          "confirmed": {
            "type": "boolean"
          }
        }
      },
      "DeviceCodePollRequest": {
        "type": "object",
        "properties": {
          "device_code": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "DeviceCodePollResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "key": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PublicApiKeyRow"
              }
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
      "DeviceCodeStartRequest": {
        "type": "object",
        "properties": {
          "client_info": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeviceCodeClientInfo"
              }
            ]
          }
        }
      },
      "DeviceCodeStartResponse": {
        "type": "object",
        "required": [
          "device_code",
          "user_code",
          "verification_uri",
          "verification_uri_complete",
          "expires_in",
          "interval"
        ],
        "properties": {
          "device_code": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64"
          },
          "interval": {
            "type": "integer",
            "format": "int64"
          },
          "user_code": {
            "type": "string"
          },
          "verification_uri": {
            "type": "string"
          },
          "verification_uri_complete": {
            "type": "string"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "MembershipRow": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "user_id",
          "role",
          "status",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "OnboardingApiKey": {
        "type": "object",
        "required": [
          "plaintext",
          "prefix",
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plaintext": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          }
        }
      },
...
      "OrganizationRow": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "plan_tier",
          "account_type",
          "seat_limit",
          "created_at"
        ],
        "properties": {
          "account_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "plan_tier": {
            "type": "string"
          },
          "seat_limit": {
            "type": "integer",
            "format": "int32"
          },
          "slug": {
            "type": "string"
          },
          "tenant_routing_tier": {
            "type": "string",
            "description": "Routing tier for this org's ClickHouse data plane.\n`\"shared\"` — routes to the shared cell (free/personal orgs).\n`\"dedicated\"` — routes to a per-org provisioned service.\nOlder records that pre-date this field deserialize to `\"dedicated\"`\n(the safe fallback) via the serde default."
          }
        }
      },
...
      "ProvisioningStatusPayload": {
        "type": "object",
        "required": [
          "status",
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string"
          },
          "service_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "SwitchOrganizationRequest": {
        "type": "object",
        "properties": {
          "org_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "UserRow": {
        "type": "object",
        "required": [
          "id",
          "primary_email",
          "created_at"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "primary_email": {
            "type": "string"
          }
        }
      },
      "UserSessionRow": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "org_id",
          "metadata",
          "created_at",
          "expires_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "InstantML SDK API key sent as Authorization: Bearer instantml_..."
      },
      "bootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-InstantML-Bootstrap-Token",
        "description": "Operator-only bootstrap token for initial users, orgs, and admin key paths"
      },
      "browserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "instantml_session",
        "description": "HttpOnly browser session cookie issued by /api/auth/dev/google or /api/auth/clerk"
      }
    }
  },
  "security": [
    {
      "bearerApiKey": []
    },
    {
      "browserSession": []
    }
  ],
...
}